Move GenerateStellarWallet to SupportMnemonic class, adding a callback
